+++ src/services/color/visualization_colors.ts
@@ -2,6 +2,6 @@
 // spots that need a large range of varied, qualitative colors.
 import { palettes } from '../../services/color/eui_palettes';
 
-export const VISUALIZATION_COLORS = palettes.euiPaletteColorBlind().colors;
+export const VISUALIZATION_COLORS = palettes.euiPaletteColorBlind();
 
 export const DEFAULT_VISUALIZATION_COLOR = VISUALIZATION_COLORS[1];